api added update

+++ cartel/cartel/src/main/java/com/project/cartel/entity/DisplayMachine.java
@@ -21,7 +21,7 @@ public class DisplayMachine {
     private String machine_name;
     private Float price;
     private int valid_days;
-    private int interest_per_day;
+    private double interest_per_day;
     private String url;
     private String description;
 }
